fix crash on initializing colorschemes

+++ drw.c
@@ -174,7 +174,7 @@ void drw_clr_create(Drw *drw, Clr *dest, const char *clrname) {
 /* Wrapper to create color schemes. The caller has to call free(3) on the
  * returned color scheme when done using it. */
 Clr *drw_scm_create(Drw *drw, const char *clrnames[], size_t clrcount) {
-    size_t i;
+    int i;
     Clr *ret;
 
     /* need at least two colors for a scheme */
